Refactor nss_simpleauthd crate
